Change data type int to Long

+++ src/main/java/com/codegym/airbnb/model/repository/HomeRepository.java
@@ -3,5 +3,5 @@ package com.codegym.airbnb.model.repository;
 import com.codegym.airbnb.model.entities.Home;
 import org.springframework.data.repository.CrudRepository;
 
-public interface HomeRepository extends CrudRepository<Home, Integer> {
+public interface HomeRepository extends CrudRepository<Home, Long> {
 }
